refactor(app): batch4 — extract useWorkspaceForgeData (ordered before editor controller)
